refactor(conflict): modularize the merge editor
