+++ lua/codediff/ui/conflict/merge/alignment.lua
@@ -153,7 +153,7 @@ local function split_up_common_equal_range_mappings(equal_ranges_1, equal_ranges
 end
 
 -- Get alignments - exact port of VSCode's getAlignments
-local function get_alignments(base_start, base_end, input1_start, input1_end, input2_start, input2_end, input1_inner_diffs, input2_inner_diffs)
+function M.get_alignments(base_start, base_end, input1_start, input1_end, input2_start, input2_end, input1_inner_diffs, input2_inner_diffs)
   -- Get equal range mappings for both inputs
   local equal_ranges_1 = to_equal_range_mappings(input1_inner_diffs, base_start, base_end, input1_start, input1_end)
   local equal_ranges_2 = to_equal_range_mappings(input2_inner_diffs, base_start, base_end, input2_start, input2_end)
@@ -232,7 +232,7 @@ end
 
 -- Exact port of VSCode's MappingAlignment.compute
 -- Takes changes from base->input1 and base->input2, returns aligned groups
-local function compute_mapping_alignments(changes1, changes2)
+function M.compute_mapping_alignments(changes1, changes2)
   -- Combine and sort all changes by base start line
   local combined = {}
   for _, c in ipairs(changes1 or {}) do
@@ -360,216 +360,6 @@ local function compute_mapping_alignments(changes1, changes2)
   push_and_reset()
 
   return alignments
-end
-
-function M.compute_merge_fillers(base_to_input1_diff, base_to_input2_diff, base_lines, input1_lines, input2_lines)
-  -- Use VSCode's MappingAlignment.compute approach
-  local mapping_alignments = compute_mapping_alignments(base_to_input1_diff.changes, base_to_input2_diff.changes)
-
-  local all_left_fillers = {}
-  local all_right_fillers = {}
-  local left_total = 0
-  local right_total = 0
-
-  for _, ma in ipairs(mapping_alignments) do
-    -- Get line alignments using VSCode's getAlignments
-    local alignments = get_alignments(
-      ma.base_range.start_line,
-      ma.base_range.end_line,
-      ma.output1_range.start_line,
-      ma.output1_range.end_line,
-      ma.output2_range.start_line,
-      ma.output2_range.end_line,
-      ma.inner1,
-      ma.inner2
-    )
-
-    -- Convert alignments to fillers
-    for _, a in ipairs(alignments) do
-      if a.input1_line and a.input2_line then
-        local left_adj = a.input1_line + left_total
-        local right_adj = a.input2_line + right_total
-        local mx = math.max(left_adj, right_adj)
-
-        if mx - left_adj > 0 then
-          table.insert(all_left_fillers, { after_line = a.input1_line - 1, count = mx - left_adj })
-          left_total = left_total + (mx - left_adj)
-        end
-        if mx - right_adj > 0 then
-          table.insert(all_right_fillers, { after_line = a.input2_line - 1, count = mx - right_adj })
-          right_total = right_total + (mx - right_adj)
-        end
-      end
-    end
-  end
-
-  return all_left_fillers, all_right_fillers
-end
-
--- Compute fillers AND identify which changes are in conflict regions
--- A conflict region is where BOTH input1 and input2 have changes to the same base region
--- This matches VSCode's behavior of only highlighting conflicting changes
-function M.compute_merge_fillers_and_conflicts(base_to_input1_diff, base_to_input2_diff, base_lines, input1_lines, input2_lines)
-  -- Use VSCode's MappingAlignment.compute approach
-  local mapping_alignments = compute_mapping_alignments(base_to_input1_diff.changes, base_to_input2_diff.changes)
-
-  local all_left_fillers = {}
-  local all_right_fillers = {}
-  local left_total = 0
-  local right_total = 0
-
-  -- Track which changes are in conflict regions
-  local conflict_left_changes = {}
-  local conflict_right_changes = {}
-
-  -- Track conflict blocks (for accept/reject actions)
-  local conflict_blocks = {}
-
-  for _, ma in ipairs(mapping_alignments) do
-    -- A region is conflicting if BOTH sides have changes (inner1 and inner2 both non-empty)
-    -- OR if both sides have changes (checked by looking at if the ranges differ from base)
-    local has_left_changes = #ma.inner1 > 0 or (ma.output1_range.end_line - ma.output1_range.start_line) ~= (ma.base_range.end_line - ma.base_range.start_line)
-    local has_right_changes = #ma.inner2 > 0 or (ma.output2_range.end_line - ma.output2_range.start_line) ~= (ma.base_range.end_line - ma.base_range.start_line)
-    local is_conflict = has_left_changes and has_right_changes
-
-    if is_conflict then
-      -- Add to conflict blocks for action handling
-      table.insert(conflict_blocks, ma)
-
-      -- Collect the actual diff changes that fall within this alignment's base range
-      for _, change in ipairs(base_to_input1_diff.changes or {}) do
-        if change.original.start_line >= ma.base_range.start_line and change.original.end_line <= ma.base_range.end_line then
-          table.insert(conflict_left_changes, change)
-        end
-      end
-      for _, change in ipairs(base_to_input2_diff.changes or {}) do
-        if change.original.start_line >= ma.base_range.start_line and change.original.end_line <= ma.base_range.end_line then
-          table.insert(conflict_right_changes, change)
-        end
-      end
-    end
-
-    -- Get line alignments using VSCode's getAlignments
-    local alignments = get_alignments(
-      ma.base_range.start_line,
-      ma.base_range.end_line,
-      ma.output1_range.start_line,
-      ma.output1_range.end_line,
-      ma.output2_range.start_line,
-      ma.output2_range.end_line,
-      ma.inner1,
-      ma.inner2
-    )
-
-    -- Convert alignments to fillers
-    for _, a in ipairs(alignments) do
-      if a.input1_line and a.input2_line then
-        local left_adj = a.input1_line + left_total
-        local right_adj = a.input2_line + right_total
-        local mx = math.max(left_adj, right_adj)
-
-        if mx - left_adj > 0 then
-          table.insert(all_left_fillers, { after_line = a.input1_line - 1, count = mx - left_adj })
-          left_total = left_total + (mx - left_adj)
-        end
-        if mx - right_adj > 0 then
-          table.insert(all_right_fillers, { after_line = a.input2_line - 1, count = mx - right_adj })
-          right_total = right_total + (mx - right_adj)
-        end
-      end
-    end
-  end
-
-  return {
-    left_fillers = all_left_fillers,
-    right_fillers = all_right_fillers,
-    conflict_blocks = conflict_blocks,
-  }, conflict_left_changes, conflict_right_changes
-end
-
---- Compute the auto-merged Result buffer content.
---- Ports VSCode's MergeEditorModel.computeAutoMergedResult().
---- For each base-range group:
----   * only input1 changed -> take input1 lines
----   * only input2 changed -> take input2 lines
----   * both changed identically -> take input1 lines
----   * both changed differently (conflict) -> keep base lines (user resolves)
---- Lines outside any group are copied straight from base.
----
---- Also returns conflict_blocks with `result_range` (1-based, inclusive start /
---- exclusive end) marking where each unresolved conflict lives in the merged
---- buffer, so extmark tracking can anchor to merged-content line numbers
---- instead of pure-base line numbers.
----
---- @param base_to_input1_diff table { changes = {...} }
---- @param base_to_input2_diff table { changes = {...} }
---- @param base_lines string[]
---- @param input1_lines string[]
---- @param input2_lines string[]
---- @return string[] merged_lines, table[] conflict_blocks
-function M.compute_auto_merged_result(base_to_input1_diff, base_to_input2_diff, base_lines, input1_lines, input2_lines)
-  local mapping_alignments = compute_mapping_alignments(base_to_input1_diff.changes, base_to_input2_diff.changes)
-
-  local result_lines = {}
-  local conflict_blocks = {}
-
-  local function append_range(source, start_line, end_line_exclusive)
-    for i = start_line, end_line_exclusive - 1 do
-      table.insert(result_lines, source[i] or "")
-    end
-  end
-
-  local function ranges_equal_content(s1, r1, s2, r2)
-    local len1 = r1.end_line - r1.start_line
-    local len2 = r2.end_line - r2.start_line
-    if len1 ~= len2 then
-      return false
-    end
-    for i = 0, len1 - 1 do
-      if (s1[r1.start_line + i] or "") ~= (s2[r2.start_line + i] or "") then
-        return false
-      end
-    end
-    return true
-  end
-
-  local base_cursor = 1 -- 1-based line index into base_lines, inclusive
-
-  for _, ma in ipairs(mapping_alignments) do
-    -- Copy unchanged base lines preceding this group.
-    append_range(base_lines, base_cursor, ma.base_range.start_line)
-    base_cursor = ma.base_range.end_line
-
-    local has1 = ma.has_input1
-    local has2 = ma.has_input2
-
-    if has1 and not has2 then
-      append_range(input1_lines, ma.output1_range.start_line, ma.output1_range.end_line)
-    elseif has2 and not has1 then
-      append_range(input2_lines, ma.output2_range.start_line, ma.output2_range.end_line)
-    elseif has1 and has2 and ranges_equal_content(input1_lines, ma.output1_range, input2_lines, ma.output2_range) then
-      -- Both sides made the identical change -> not a conflict, apply once.
-      append_range(input1_lines, ma.output1_range.start_line, ma.output1_range.end_line)
-    else
-      -- True conflict: keep base, leave for the user to resolve.
-      local conflict_start = #result_lines + 1
-      append_range(base_lines, ma.base_range.start_line, ma.base_range.end_line)
-      local conflict_end_exclusive = #result_lines + 1
-      table.insert(conflict_blocks, {
-        base_range = ma.base_range,
-        output1_range = ma.output1_range,
-        output2_range = ma.output2_range,
-        inner1 = ma.inner1,
-        inner2 = ma.inner2,
-        result_range = { start_line = conflict_start, end_line = conflict_end_exclusive },
-      })
-    end
-  end
-
-  -- Copy any remaining base lines after the last group.
-  append_range(base_lines, base_cursor, #base_lines + 1)
-
-  return result_lines, conflict_blocks
 end
 
 return M
