+++ lua/codediff/ui/view/side_by_side/update.lua
@@ -9,8 +9,8 @@ local layout = require("codediff.ui.layout")
 local helpers = require("codediff.ui.view.helpers")
 local readiness = require("codediff.ui.view.readiness")
 local render = require("codediff.ui.view.render")
+local conflict_view = require("codediff.ui.conflict.view")
 local view_keymaps = require("codediff.ui.view.keymaps")
-local conflict_window = require("codediff.ui.view.conflict_window")
 local welcome_window = require("codediff.ui.view.welcome_window")
 
 local is_virtual_revision = helpers.is_virtual_revision
@@ -18,9 +18,9 @@ local prepare_buffer = helpers.prepare_buffer
 local show_real_file_buffer = helpers.show_real_file_buffer
 local open_real_file = helpers.open_real_file
 local compute_and_render = render.compute_and_render
-local compute_and_render_conflict = render.compute_and_render_conflict
+local compute_and_render_conflict = conflict_view.compute_and_render_conflict
 local setup_auto_refresh = render.setup_auto_refresh
-local setup_conflict_result_window = conflict_window.setup_conflict_result_window
+local setup_conflict_result_window = conflict_view.setup_conflict_result_window
 local setup_all_keymaps = view_keymaps.setup_all_keymaps
 
 --- Put one side's content into `win` during an update, reusing the buffer when
