+++ tests/module_loading_spec.lua
@@ -273,9 +273,21 @@ end)
 -- ── conflict ────────────────────────────────────────────────────────────
 describe("Conflict submodules", function()
   describe("module loading", function()
-    it("loads actions module", function()
-      local ok, mod = pcall(require, "codediff.ui.conflict.actions")
-      assert.is_true(ok, "Failed to require codediff.ui.conflict.actions")
+    it("loads merge module", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict.merge")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict.merge")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads view module", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict.view")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict.view")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads resolution module", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict.resolution")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict.resolution")
       assert.is_not_nil(mod)
     end)
 
@@ -297,12 +309,6 @@ describe("Conflict submodules", function()
       assert.is_not_nil(mod)
     end)
 
-    it("loads diffget module", function()
-      local ok, mod = pcall(require, "codediff.ui.conflict.diffget")
-      assert.is_true(ok, "Failed to require codediff.ui.conflict.diffget")
-      assert.is_not_nil(mod)
-    end)
-
     it("loads init facade", function()
       local ok, mod = pcall(require, "codediff.ui.conflict")
       assert.is_true(ok, "Failed to require codediff.ui.conflict")
@@ -311,8 +317,33 @@ describe("Conflict submodules", function()
   end)
 
   describe("public API", function()
-    it("actions exports expected functions", function()
-      local mod = require("codediff.ui.conflict.actions")
+    it("merge facade delegates expected functions", function()
+      local mod = require("codediff.ui.conflict.merge")
+      local fillers = require("codediff.ui.conflict.merge.fillers")
+      local auto_merge = require("codediff.ui.conflict.merge.auto_merge")
+      assert.is_function(mod.compute_merge_fillers)
+      assert.is_function(mod.compute_merge_fillers_and_conflicts)
+      assert.is_function(mod.compute_auto_merged_result)
+      assert.equal(fillers.compute_merge_fillers, mod.compute_merge_fillers)
+      assert.equal(fillers.compute_merge_fillers_and_conflicts, mod.compute_merge_fillers_and_conflicts)
+      assert.equal(auto_merge.compute_auto_merged_result, mod.compute_auto_merged_result)
+    end)
+
+    it("view facade delegates expected functions", function()
+      local mod = require("codediff.ui.conflict.view")
+      local inputs = require("codediff.ui.conflict.view.inputs")
+      local result = require("codediff.ui.conflict.view.result")
+      assert.is_function(mod.compute_and_render_conflict)
+      assert.is_function(mod.setup_conflict_result_window)
+      assert.equal(inputs.compute_and_render_conflict, mod.compute_and_render_conflict)
+      assert.equal(result.setup_conflict_result_window, mod.setup_conflict_result_window)
+    end)
+
+    it("resolution facade delegates expected functions", function()
+      local mod = require("codediff.ui.conflict.resolution")
+      local block = require("codediff.ui.conflict.resolution.block")
+      local file = require("codediff.ui.conflict.resolution.file")
+      local diffget = require("codediff.ui.conflict.resolution.diffget")
       assert.is_function(mod.accept_incoming)
       assert.is_function(mod.accept_current)
       assert.is_function(mod.accept_both)
@@ -321,6 +352,18 @@ describe("Conflict submodules", function()
       assert.is_function(mod.accept_all_current)
       assert.is_function(mod.accept_all_both)
       assert.is_function(mod.discard_all)
+      assert.is_function(mod.diffget_incoming)
+      assert.is_function(mod.diffget_current)
+      assert.equal(block.accept_incoming, mod.accept_incoming)
+      assert.equal(block.accept_current, mod.accept_current)
+      assert.equal(block.accept_both, mod.accept_both)
+      assert.equal(block.discard, mod.discard)
+      assert.equal(file.accept_all_incoming, mod.accept_all_incoming)
+      assert.equal(file.accept_all_current, mod.accept_all_current)
+      assert.equal(file.accept_all_both, mod.accept_all_both)
+      assert.equal(file.discard_all, mod.discard_all)
+      assert.equal(diffget.diffget_incoming, mod.diffget_incoming)
+      assert.equal(diffget.diffget_current, mod.diffget_current)
     end)
 
     it("keymaps exports setup function", function()
@@ -340,10 +383,35 @@ describe("Conflict submodules", function()
       assert.is_function(mod.setup_sign_refresh_autocmd)
     end)
 
-    it("diffget exports expected functions", function()
-      local mod = require("codediff.ui.conflict.diffget")
+    it("init facade delegates expected functions", function()
+      local mod = require("codediff.ui.conflict")
+      local tracking = require("codediff.ui.conflict.tracking")
+      local signs = require("codediff.ui.conflict.signs")
+      local resolution = require("codediff.ui.conflict.resolution")
+      local navigation = require("codediff.ui.conflict.navigation")
+      local keymaps = require("codediff.ui.conflict.keymaps")
+      assert.is_function(mod.run_repeatable_action)
+      assert.is_function(mod.initialize_tracking)
+      assert.is_function(mod.accept_incoming)
       assert.is_function(mod.diffget_incoming)
-      assert.is_function(mod.diffget_current)
+      assert.is_function(mod.navigate_next_conflict)
+      assert.equal(tracking.run_repeatable_action, mod.run_repeatable_action)
+      assert.equal(tracking.initialize_tracking, mod.initialize_tracking)
+      assert.equal(signs.refresh_all_conflict_signs, mod.refresh_all_conflict_signs)
+      assert.equal(signs.setup_sign_refresh_autocmd, mod.setup_sign_refresh_autocmd)
+      assert.equal(resolution.accept_incoming, mod.accept_incoming)
+      assert.equal(resolution.accept_current, mod.accept_current)
+      assert.equal(resolution.accept_both, mod.accept_both)
+      assert.equal(resolution.discard, mod.discard)
+      assert.equal(resolution.accept_all_incoming, mod.accept_all_incoming)
+      assert.equal(resolution.accept_all_current, mod.accept_all_current)
+      assert.equal(resolution.accept_all_both, mod.accept_all_both)
+      assert.equal(resolution.discard_all, mod.discard_all)
+      assert.equal(resolution.diffget_incoming, mod.diffget_incoming)
+      assert.equal(resolution.diffget_current, mod.diffget_current)
+      assert.equal(navigation.navigate_next_conflict, mod.navigate_next_conflict)
+      assert.equal(navigation.navigate_prev_conflict, mod.navigate_prev_conflict)
+      assert.equal(keymaps.setup_keymaps, mod.setup_keymaps)
     end)
   end)
 end)
